requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
pillow==10.2.0
boto3==1.34.34
python-decouple==3.8
django-storages==1.14.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0